requirements.txt
langchain==0.3.26
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langchain-community==0.3.27
faiss-cpu>=1.7.0
unstructured==0.18.1
Markdown==3.8.2
sentence-transformers>=3.0.0
lazy_loader==0.4
rank_bm25==0.2.2
openai>=1.86.0,<2.0.0

# RagFlow支持多格式文件处理的额外依赖
ragflow>=0.0.1

# PDF处理
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Word文档处理
python-docx>=0.8.11

# Excel文件处理
openpyxl>=3.1.0
xlsxwriter>=3.1.9

# CSV和数据表格处理
pandas>=2.0.0

# 图片OCR处理
Pillow>=10.0.0
pytesseract>=0.3.10

# 其他文本格式
chardet>=5.0.0

# 网页处理
beautifulsoup4>=4.12.0
lxml>=4.9.0